ui redesign polishes
